+++ include/adm/utilities/block_duration_assignment.hpp
@@ -4,6 +4,7 @@
 #include <chrono>
 #include <memory>
 #include "adm/export.h"
+#include "adm/elements/time.hpp"
 
 namespace adm {
 
@@ -33,8 +34,8 @@ namespace adm {
    * @param fileLength The length of the BW64 audio file
    * @sa void updateBlockFormatDurations(std::shared_ptr<Document>)
    */
-  ADM_EXPORT void updateBlockFormatDurations(
-      std::shared_ptr<Document> document, std::chrono::nanoseconds fileLength);
+  ADM_EXPORT void updateBlockFormatDurations(std::shared_ptr<Document> document,
+                                             const Time& fileLength);
 
   /**
    * @brief Set or update durations of all `AudioBlockFormats`